tests/wpt/tests/css/css-overflow/scrollable-overflow-iframe-uncontained-fixed-positioned.html

<!DOCTYPE html>
<html>
<head>
<title>Scrollable overflow of iframe's document element with uncontained fixed position</title>
<link rel="help" href="https://drafts.csswg.org/css-overflow/#scrollable">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-element-scrollwidth">
<link rel="help" href="https://drafts.csswg.org/cssom-view/#dom-element-scrollheight">
<link rel="author" title="Jo Steven Novaryo" href="mailto:jo.steven.novaryo@huawei.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <iframe
    id="target"
    srcdoc="<!DOCTYPE html> <div style='position: fixed; width: 1000px; height: 1000px'></div>"
    width="100"
    height="100"
  >
  </iframe>
</body>
<script>
promise_test(async (t) => {
  await new Promise(resolve => target.addEventListener('load', resolve));

  assert_less_than(target.contentDocument.documentElement.scrollHeight, 1000);
  assert_less_than(target.contentDocument.documentElement.scrollWidth, 1000);

}, "Uncontained fixed positioned shouldn't affect scrollable overflow of a document element.");
</script>
</html>
